+++ src/main/java/ru/kata/spring/boot_security/demo/repository/RoleRepositoryImpl.java
@@ -20,7 +20,7 @@ public class RoleRepositoryImpl implements RoleRepository {
     }
 
     @Override
-    public Role findRole(Integer id) {
+    public Role findRole(Long id) {
         return entityManager.find(Role.class, id);
     }
